+++ src/utils/types.ts
@@ -116,3 +116,12 @@ export function isObject(input: any): input is object {
 export function isArray(input: any): input is any[] {
   return input instanceof Array;
 }
+
+export function arrayEquals(a: any[], b: any[]) {
+  return (
+    Array.isArray(a) &&
+    Array.isArray(b) &&
+    a.length === b.length &&
+    a.every((val, index) => val === b[index])
+  );
+}